redo timing patcher
upgraded to new version of the 'timing patcher' (not included in the repo) and
reincorporated the stop/start looping functionality. Moved the set
looping code from the foot pedal patcher to the timing patcher. Kept a
snippet of the code to add looping functionality to the timing patcher
in case it is lost or something