change so all foot pedal info is sent for fishman
parsing of 99 options from foot pedal happens within the fishman
patches. need to extend this mode of operation to others now that I have
figured out how to reliably have a dependency throughout the library.
this particular change updates the fishman pach identifier when the
update track info button is pressed (when track are moved around)